[PRODDEV-111] Refactoring to avoid the duplicated code
