+++ modules/openy_gc_auth/modules/openy_gc_auth_yusa/src/Controller/UserController.php
@@ -6,6 +6,7 @@ use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Controller\ControllerBase;
 use Drupal\Core\Flood\FloodInterface;
 use Drupal\openy_gc_auth\GCUserAuthorizer;
+use Drupal\openy_gc_auth\GCVerificationTrait;
 use Drupal\openy_gc_auth_yusa\YUSAClientService;
 use Drupal\user\Entity\User;
 use Drupal\user\UserDataInterface;
@@ -21,6 +22,8 @@ use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
  * Controller routines for user routes.
  */
 class UserController extends ControllerBase {
+
+  use GCVerificationTrait;
 
   /**
    * The user storage.
@@ -181,13 +184,7 @@ class UserController extends ControllerBase {
       else {
         $id = $email;
       }
-      $verified_browsers = [];
-      if (!$request->cookies->has('Drupal_visitor_auth_yusa_authorized')) {
-        $verified_browsers = $this->userData->get('openy_gc_auth_yusa', $user->id(), 'verified_browsers');
-      }
-      $token = user_pass_rehash($user, $current);
-      $verified_browsers[$token] = $current;
-      $this->userData->set('openy_gc_auth_yusa', $user->id(), 'verified_browsers', $verified_browsers);
+      $token = $this->saveVerification($request, $user, $current);
 
       $this
         ->gcUserAuthorizer
@@ -199,7 +196,7 @@ class UserController extends ControllerBase {
       // Clear any flood events for this IP.
       $this->flood->clear('openy_gc_auth_yusa.login');
       $response = new RedirectResponse($vy_settings->get('virtual_y_url'), 302);
-      $response->headers->setCookie(new Cookie('Drupal_visitor_auth_yusa_authorized', $token));
+      $response->headers->setCookie(new Cookie('Drupal_visitor_gc_auth_authorized', $token));
       return $response;
     }
